+++ src/servant/tt_002/6C98.c
@@ -1,0 +1,4 @@
+// SPDX-License-Identifier: AGPL-3.0-or-later
+#include "faerie.h"
+
+INCLUDE_ASM("servant/tt_002/nonmatchings/6C98", func_us_80176C98);
